Refactor a bit and fix sharing on Android.
