+++ src/screens/SharedRecordsScreen.js
@@ -28,42 +28,7 @@ class SharedRecordsScreenUnwrapped extends Component {
     let sharedRecords = await store.sharedRecords();
     let externalRecords = await store.externalRecords();
 
-    let contactsArray = [];
-
-    for (const [email, records] of Object.entries(sharedRecords)) { 
-      let contact = contacts.find(function (c) { return c.name === email; });
-
-      if (contact) {
-        let recordArray = [];
-
-        contact.key = contact.name;
-
-        for (const [id, record] of Object.entries(records)) {
-          recordArray.push({id, record});
-        }            
-
-        contactsArray.push({contact, records: recordArray});
-      }
-    }
-
-    for (const [email, records] of Object.entries(externalRecords)) { 
-      let contact = null;
-      let recordArray = [];
-
-      for (const [id, record] of Object.entries(records)) {
-        if (!contact) {
-          contact = record.externalContact;
-          contact.key = `_${contact.name}`;
-          contact.external = true;
-        }
-
-        recordArray.push({id, record});
-      }
-
-      if (recordArray.length > 0) {
-        contactsArray.push({contact, records: recordArray});
-      }
-    }
+    let contactsArray = importHelpers.groupByContact(contacts, sharedRecords, externalRecords);
 
     this.setState({ contacts: contactsArray, loading: false });
   }
@@ -108,39 +73,7 @@ class SharedRecordsScreenUnwrapped extends Component {
    * we have in state to clean-up or add new entries.
    */
   mergeContacts(local, updates) {
-    let result = local;
-
-    for (update of updates.removed) {
-      let key = (update.contact.external) ? `_${update.contact.name}` : update.contact.name;
-
-      let entry = result.find(function (entry) { return entry.contact.key === key; });
-
-      if (entry) {
-        let entryRecords = entry.records.filter((r) => (r.id != update.record.id));
-        if (entryRecords.length === 0) {
-          result = result.filter((entry) => (entry.contact.key !== key));
-        } else {
-          entry.records = entryRecords;
-        }
-      }
-    }
-
-    for (update of updates.added) {
-      let key = (update.contact.external) ? `_${update.contact.name}` : update.contact.name;
-
-      let entry = result.find(function (entry) { return entry.contact.key === key; });
-
-      if (entry) {
-        let record = entry.records.find(function (r) { return r.id === update.record.id; });
-
-        if (!record) {
-          entry.records.push({id: update.record.id, record: update.record});
-        }
-      } else {
-        update.contact.key = key;
-        result.push({contact: update.contact, records: [{id:update.record.id, record:update.record}]});
-      }      
-    }
+    let result = importHelpers.applySharedRecordUpdates(local, updates);
 
     // We don't se setState on purpose. This is just to update the state variable with the updates. The result will already
     // supply the correct array to the render method.
